# bib_api/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Bib API",
    "description": "Dokumentation für alle Endpunkte der Bibliotheks API",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "301": {
            "description": "Moved Permanently"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to login a user.",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "Details of the user to add.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "desk@library.example"
                },
                "password": {
                  "type": "string",
                  "example": "m295"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/verify": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to verify if the user is logged in.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/logout": {
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to logout a user.",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/books/books": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "Endpoint to get the list of books.",
        "parameters": [
          {
            "name": "isbn",
            "in": "query",
            "description": "ISBN of the book to search for.",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Book"
        ],
        "description": "Endpoint to add a new book.",
        "parameters": [
          {
            "name": "book",
            "in": "body",
            "description": "Details of the book to add.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "isbn": {
                  "type": "string",
                  "example": "978-3-16-148410-0"
                },
                "title": {
                  "type": "string",
                  "example": "The Great Gatsby"
                },
                "author": {
                  "type": "string",
                  "example": "F. Scott Fitzgerald"
                },
                "publisher": {
                  "type": "string",
                  "example": "Scribner"
                },
                "year": {
                  "type": "number",
                  "example": 1925
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/books/books/{isbn}": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "Endpoint to get a book by ISBN.",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ISBN of the book."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Book"
        ],
        "description": "Endpoint to update a book.",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ISBN of the book."
          },
          {
            "name": "book",
            "in": "body",
            "description": "Updated details of the book.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "isbn": {
                  "type": "string",
                  "example": "9781234567890"
                },
                "title": {
                  "type": "string",
                  "example": "idk"
                },
                "author": {
                  "type": "string",
                  "example": "molmol"
                },
                "publisher": {
                  "type": "string",
                  "example": "Unknown Publisher"
                },
                "year": {
                  "type": "number",
                  "example": 2069
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Book"
        ],
        "description": "Endpoint to delete a book by ISBN.",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ISBN of the book."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Book"
        ],
        "description": "Endpoint to partially update a book.",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ISBN of the book."
          },
          {
            "name": "book",
            "in": "body",
            "description": "Details of the book to update.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "The bad Gatsby"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/lends/lends": {
      "get": {
        "tags": [
          "Lend"
        ],
        "description": "Endpoint to get the list of lends.",
        "parameters": [
          {
            "name": "isbn",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Lend"
        ],
        "description": "Endpoint to add a new lend.",
        "parameters": [
          {
            "name": "lend",
            "in": "body",
            "description": "Details of the lend to add.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "customer_id": {
                  "type": "number",
                  "example": 1
                },
                "isbn": {
                  "type": "string",
                  "example": "724895934-0"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/lends/lends/{isbn}": {
      "get": {
        "tags": [
          "Lend"
        ],
        "description": "Endpoint to get a lend by ISBN.",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ISBN of the lend."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Lend"
        ],
        "description": "Endpoint to delete a lend by ISBN.",
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "lend",
            "in": "body",
            "description": "Details of the lend to add.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "customer_id": {
                  "type": "number",
                  "example": 1
                },
                "isbn": {
                  "type": "string",
                  "example": "724895934-0"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}